% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_schemes.R
\docType{data}
\name{classification_schemes}
\alias{classification_schemes}
\title{Data on classification schemes}
\format{
A data frame with 10 variables:
\describe{
\item{key_id}{Numeric. An ID number that uniquely identifies each
observation. Indicates the default sort order for the dataset.}
\item{scheme_id}{Numeric. An ID number that uniquely identifies each
classification scheme. Coded \code{1} for the pre-Lisbon version of
EUR-Lex's Case Law Directory, coded \code{2} for the post-Lisbon version of
EUR-Lex's Case Law Directory, and coded \code{3} for EUR-Lex's Directory of
Legal Acts. See the variable \code{scheme}.}
\item{scheme}{String. The classification scheme. Possible values include:
\code{Pre-Lisbon version of EUR-Lex's Case Law Directory},
\code{Post-Lisbon version of EUR-Lex's Case Law Directory}, and
\code{EUR-Lex's Directory of Legal Acts}.}
\item{code_id}{Numeric. An ID number that uniquely identifies each
combination of major and minor code within each scheme. }
\item{major_code}{String. The major code.}
\item{major_code_topic}{String. The description of the major code.}
\item{minor_code}{String. The minor code. Coded \code{None} if there are no
minor codes that correspond to the major code.}
\item{minor_code_topic}{String. The description of the minor code. Coded
\code{None} if there are no minor codes that correspond to the major code.}
\item{policy_area_id}{Numeric. An ID number that uniquely identifies each
policy area. See the \code{policy_areas} dataset.}
\item{policy_area}{String. The name of the policy area. See the
\code{policy_areas} dataset.}
}
}
\usage{
classification_schemes
}
\description{
This dataset includes data on the classification schemes used by EUR-Lex,
the Commission's official database of EU legal documents. EUR-Lex sorts
documents based on a case law directory and a directory of legal acts.
There is a pre-Lisbon and post-Lisbon version of the case law directory.
Each directory has major codes and minor codes. There is one observation
per minor code per classification scheme. The dataset indicates the policy
area (if any) that best corresponds to each minor code.
}
\keyword{datasets}
