% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departments_by_commission.R
\docType{data}
\name{departments_by_commission}
\alias{departments_by_commission}
\title{Data on departments by commission}
\format{
A data frame with 18 variables:
\describe{
\item{key_id}{\code{numeric}. An ID number that uniquely identifies each
observation. Indicates the default sort order for the dataset.}
\item{commission_id}{\code{numeric}. The ID number that uniquely identifies
each Commission. See the \code{commissions} dataset.}
\item{commission}{\code{string}. The name of the Commission. See the
\code{commissions} dataset.}
\item{start_date}{\code{date}. The start date of the Commission.}
\item{start_year}{\code{numeric}. The start year of the Commission.}
\item{start_month}{\code{numeric}. The start month of the Commission.}
\item{start_day}{\code{numeric}. The start day of the Commission.}
\item{end_date}{\code{date}. The end date of the Commission.}
\item{end_year}{\code{numeric}. The start year of the Commission.}
\item{end_month}{\code{numeric}. The start month of the Commission.}
\item{end_day}{\code{numeric}. The end day of the Commission.}
\item{department_type_id}{\code{numeric}. An ID number that uniquely
identifies the type of the department. Coded \code{1} for
\code{Directorate-General} and \code{2} for \code{Service department}. See
the variable \code{department_type}.}
\item{department_type}{\code{string}. The type of the department. Either
\code{Directorate-General} or \code{Service department}.}
\item{department_id}{\code{numeric}. An ID number that uniquely identifies
each department. See the \code{departments} dataset.}
\item{department}{\code{string}. The name of the department.}
\item{department_code}{\code{string}. The common abbreviation of the
department. Coded \code{Not applicable} of there is not an abbreviation for
the department. Coded \code{None} for all Commissions prior to the Prodi
Commission.}
\item{policy_area}{\code{string}. The name of the policy area. See the
\code{policy_areas} dataset.}
\item{current_department}{\code{string}. The name of the current
department. See the \code{departments} dataset.}
}
}
\usage{
departments_by_commission
}
\description{
This dataset includes data on departments of the Commission, including
Directorates-General and service departments (1958-2021). There is one
observation per department per Commission. This dataset provides a snapshot
of each Commission at the time that the first directory is published. For
Commissions that come into office early in a year, the first directory is
usually published later that year. For Commissions that come into office
late in a year, the first directory is usually published early the next
year. The exception is the Prodi Commission. The data for this Commission
comes from the earliest available directory.
}
\keyword{datasets}
