% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_area_histories.R
\docType{data}
\name{policy_area_histories}
\alias{policy_area_histories}
\title{Data on policy area histories}
\format{
A data frame with 18 variables:
\describe{
\item{key_id}{Numeric. An ID number that uniquely identifies each
observation. Indicates the default sort order for the dataset.}
\item{policy_area_id}{Numeric. An ID number that uniquely identifies each
policy area. See the \code{policy_areas} dataset.}
\item{policy_area}{String. The name of the policy area. See the
\code{policy_areas} dataset.}
\item{department_id}{Numeric. An ID number that uniquely identifies each
department. See the \code{departments} dataset.}
\item{department}{String. The name of the department.}
\item{department_code}{String. The common abbreviation of the department.
Coded \code{Not applicable} of there is not an abbreviation for the
department. Coded \code{None} for all Commissions prior to the Prodi
Commission.}
\item{department_type_id}{Numeric. An ID number that uniquely identifies
the type of the department. Coded \code{1} for \code{Directorate-General}
and \code{2} for \code{Service department}. See the variable
\code{department_type}.}
\item{department_type}{String. The type of the department. Either
\code{Directorate-General} or \code{Service department}.}
\item{commissions}{String. A list of the Commissions in which the policy
area appeared, separated by a comma.}
\item{count_commissions}{Numeric. The number of Commissions in which the
policy area appeared.}
\item{start_date}{Date. The start date of the historical interval.
Historical intervals can cross Commissions.}
\item{start_year}{Numeric. The start year of the historical interval.
Historical intervals can cross Commissions.}
\item{start_month}{Numeric. The start month of the historical interval.
Historical intervals can cross Commissions.}
\item{start_day}{Numeric. The start day of the historical interval.
Historical intervals can cross Commissions.}
\item{end_date}{Date. The end date of the historical interval. Historical
intervals can cross Commissions.}
\item{end_year}{Numeric. The start year of the historical interval.
Historical intervals can cross Commissions.}
\item{end_month}{Numeric. The start month of the historical interval.
Historical intervals can cross Commissions.}
\item{end_day}{Numeric. The end day of the historical interval. Historical
intervals can cross Commissions.}
}
}
\usage{
policy_area_histories
}
\description{
This dataset includes data on the histories of the policy areas covered by
the departments of the Commission (1958-2021). There is one observation per
historical interval in which the mapping between a policy area and a
department is constant. If the nomenclature of a department changes, a new
interval will be created. Unlike allocations, intervals can span
Commissions. There can also be multiple overlaping intervals per policy
area when a policy area is split across multiple departments at the same
time. By sorting the dataset by policy area, you can see the evolution of
the mapping between the policy area and departments over time. The dataset
indicates the Commissions associated with each interval and the start and
end dates of each interval.
}
\keyword{datasets}
