% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_areas.R
\docType{data}
\name{policy_areas}
\alias{policy_areas}
\title{Data on policy areas}
\format{
A data frame with 6 variables:
\describe{
\item{key_id}{Numeric. An ID number that uniquely identifies each
observation. Indicates the default sort order for the dataset.}
\item{policy_area_id}{Numeric. An ID number that uniquely identifies each
policy area. See the \code{policy_areas} dataset.}
\item{policy_area}{String. The name of the policy area. See the
\code{policy_areas} dataset.}
\item{case_law_old}{String. A list of the codes from the pre-Lisbon version
of EUR-Lex's Case Law Directory that apply to the policy area.}
\item{case_law_new}{String. A list of the codes from the post-Lisbon
version of EUR-Lex's Case Law Directory that apply to the policy area.}
\item{legislation}{String. A list of the codes from EUR-Lex's Directory of
Legal Acts that apply to the policy area.}
}
}
\usage{
policy_areas
}
\description{
This dataset includes data on policy areas covered by departments of the
Commission (1958-2021). There is one observation per policy area. EUR-Lex,
the Commission's official database of EU legal documents, sorts documents
based on a case law directory and a directory of legal acts. There is a
pre-Lisbon and post-Lisbon version of the case law directory. The dataset
indicates the directory codes that correspond to each policy area according
to each of these classification schemes.
}
\keyword{datasets}
