% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_allocations.R
\docType{data}
\name{portfolio_allocations}
\alias{portfolio_allocations}
\title{Data on portfolio allocations}
\format{
A data frame with 28 variables:
\describe{
\item{key_id}{Numeric. An ID number that uniquely identifies each
observation. Indicates the default sort order for the dataset.}
\item{commission_id}{Numeric. The ID number that uniquely identifies each
Commission. See the \code{commissions} dataset.}
\item{commission}{String. The name of the Commission. See the
\code{commissions} dataset.}
\item{commissioner_id}{Numeric. An ID number that uniquely identifies each
individual who was served as a Commissioner. See the \code{commissioners}
dataset.}
\item{full_name}{String. The full name of the Commissioner.}
\item{full_name_latin}{String. The full name of the Commissioner using only
basic Latin characters.}
\item{last_name}{String. The last name of the Commissioner.}
\item{last_name_latin}{String. The last name of the Commissioner using only
basic Latin characters.}
\item{first_name}{String. The first name of the Commissioner.}
\item{first_name_latin}{String. The first name of the Commissioner using
only basic Latin characters.}
\item{member_state_id}{Numeric. An ID number that uniquely identifies each
member state.}
\item{member_state}{String. The name of the Commissioner's member state.}
\item{member_state_code}{String. The two-letter code assigned by the
Commission to the Commissioner's member state.}
\item{national_party_code}{String. The common abbreviation of the
Commissioner's national political party.}
\item{national_party_name}{String. The name of the Commissioner's national
political party in the Commissioner's member state's official language.}
\item{national_party_name_english}{String. The name of the Commissioner's
national political party in English.}
\item{political_group_code}{String. The common abbreviation of the
political group.}
\item{political_group_name}{String. The name of the Commissioner's
political group.}
\item{start_date}{Date. The start date of the allocation. Allocations do
not cross Commissions.}
\item{start_year}{Numeric. The start year of the allocation. Allocations do
not cross Commissions.}
\item{start_month}{Numeric. The start month of the allocation. Allocations
do not cross Commissions.}
\item{start_day}{Numeric. The start day of the allocation. Allocations do
not cross Commissions.}
\item{end_date}{Date. The end date of the allocation. Allocations do not
cross Commissions.}
\item{end_year}{Numeric. The start year of the allocation. Allocations do
not cross Commissions.}
\item{end_month}{Numeric. The start month of the allocation. Allocations do
not cross Commissions.}
\item{end_day}{Numeric. The end day of the allocation. Allocations do not
cross Commissions.}
\item{portfolio}{String. The Commissioner's portfolio. If the portfolio has
multiple components, they are listed, separated by a semi-colon. }
\item{departments}{String. A list of the departments that the portfolio
includes, separated by a semi-colon.}
}
}
\usage{
portfolio_allocations
}
\description{
This dataset includes data on portfolio allocations (1958-2021). There is
one observation per portfolio allocation per Commission (i.e., one
observation per portfolio per Commissioner per Commission). A new
observation is created every time the portfolio of a Commissioner changes,
regardless of whether that change results in a change in the departments
that the Commissioner oversees. There can be multiple observations per
individual per Commission if an individual's portfolio changes during the
Commission. Portfolio allocations do not span Commissions. If a
Commissioner has the same portfolio for more than one Commission, there
will be multiple observations, one for each Commission. The dataset
indicates the start and end dates of each portfolio allocation, the
departments that each Commissioner oversees, and the policy areas that
correspond to those departments.
}
\keyword{datasets}
